refactor(spark): share Spark 4.x mapping and internal-row impls in hudi-spark4-common
Extract the bodies duplicated across the Spark 4.0/4.1/4.2 copies of
HoodiePartitionFileSliceMapping, HoodiePartitionCDCFileGroupMapping, and
HoodieInternalRow into shared traits/abstract classes in hudi-spark4-common,
mirroring how hudi-spark3-common already shares them for the 3.x family.
Version classes keep only genuine deltas: the Spark 4.1/4.2 geography and
geometry getters, and the concrete type returned by copy() via a
newInternalRow factory hook.
