+++ hudi-spark-datasource/hudi-spark4.1.x/src/main/scala/org/apache/hudi/Spark41HoodiePartitionCDCFileGroupMapping.scala
@@ -24,11 +24,6 @@ import org.apache.hudi.common.table.cdc.HoodieCDCFileSplit
 import org.apache.spark.sql.catalyst.InternalRow
 
 class Spark41HoodiePartitionCDCFileGroupMapping(partitionValues: InternalRow,
-                                                fileSplits: List[HoodieCDCFileSplit])
+                                                protected val fileSplits: List[HoodieCDCFileSplit])
   extends Spark41HoodiePartitionValues(partitionValues)
-    with HoodiePartitionCDCFileGroupMapping {
-
-  override def getFileSplits(): List[HoodieCDCFileSplit] = {
-    fileSplits
-  }
-}
+  with Spark4HoodiePartitionCDCFileGroupMapping
