+++ hudi-spark-datasource/hudi-spark4.1.x/src/main/scala/org/apache/hudi/Spark41HoodiePartitionFileSliceMapping.scala
@@ -24,13 +24,6 @@ import org.apache.hudi.common.model.FileSlice
 import org.apache.spark.sql.catalyst.InternalRow
 
 class Spark41HoodiePartitionFileSliceMapping(values: InternalRow,
-                                             slices: Map[String, FileSlice])
+                                             protected val slices: Map[String, FileSlice])
   extends Spark41HoodiePartitionValues(values)
-    with HoodiePartitionFileSliceMapping {
-
-  override def getSlice(fileId: String): Option[FileSlice] = {
-    slices.get(fileId)
-  }
-
-  override def getPartitionValues: InternalRow = values
-}
+  with Spark4HoodiePartitionFileSliceMapping
